Re #167 mtimesx_horace changes to work on Linux
git-svn-id: https://svn.isis.rl.ac.uk/Horace/trunk@1542 1cb955da-4463-431d-88dd-2af2f5550de3

+++ _LowLevelCode/cpp/mtimesx_horace/MatMultiply.h
@@ -18,7 +18,7 @@ size_t calc_mdims(mwSize const *const dimsA, size_t ndimsA);
 
 /* Calculate size and the dimensionality of the resulting array */
 void calc_output_size(mwSize const *const dimsA, size_t ndimsA, mwSize const *const  dimsB, size_t ndimsB,
-    std::vector<size_t> & rez_dim_sizes, size_t &nDims, size_t &Mk, bool & expandA, bool &expandB);
+    std::vector<int> & rez_dim_sizes, size_t &nDims, size_t &Mk, bool & expandA, bool &expandB);
 
 
 /* multiply two matrices or array of matrices */
